requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlmodel==0.0.22
alembic==1.13.3
psycopg2-binary==2.9.10
PyPDF2==3.0.1
python-pptx==1.0.2
openai==1.99.9
aiofiles==24.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
